Working on icache
Just saving my current work

+++ src/config.rs
@@ -1,5 +1,5 @@
 // pub type BxPhyAddress = u64;
-pub type BxPhyAddress = usize;
+pub type BxPhyAddress = u64;
 
 pub type BxAddress = u64;
